ml/requirements.txt
pandas>=2.1
numpy>=1.26
scikit-learn>=1.4
xgboost>=2.0
lightgbm>=4.3
torch>=2.2
shap>=0.45
matplotlib>=3.8
joblib>=1.3
pyyaml>=6.0
python-dotenv>=1.0
kaggle>=1.6
supabase>=2.4
tqdm>=4.66
transformers>=4.40
sentence-transformers>=2.7
pytest>=8.0
